move filter to data_set
