+++ pyvista/core/filters/poly_data.py
@@ -3540,105 +3540,6 @@ class PolyDataFilters(DataSetFilters):
             return mesh, pyvista.wrap(alg.GetContourEdgesOutput())
         return mesh
 
-    def iterative_closest_point(
-        self,
-        target,
-        max_landmarks=100,
-        max_mean_distance=1e-5,
-        max_iterations=500,
-        check_mean_distance=True,
-        start_by_matching_centroids=True,
-        return_matrix=False,
-    ):
-        """Perform iterative closest point registration between two meshes.
-
-        Parameters
-        ----------
-        target : pyvista.PolyData
-            The target mesh to align the input mesh.
-
-        max_landmarks : int, default: 100
-            The maximum number of landmarks.
-
-        max_mean_distance : float, default: 1e-5
-            The maximum mean distance for convergence.
-
-        max_iterations : int, default: 500
-            The maximum number of iterations.
-
-        check_mean_distance : bool, default: True
-            Whether to check the mean distance for convergence.
-
-        start_by_matching_centroids : bool, default: True
-            Whether to start the alignment by matching centroids. Default is True.
-
-        return_matrix : bool, default: False
-            Return the transform matrix as well as the aligned mesh.
-
-        Returns
-        -------
-        aligned : pyvista.PolyData
-            The input mesh aligned to the target mesh.
-
-        matrix : numpy.ndarray
-            Transform matrix to transform the input mesh to the target.
-
-        Examples
-        --------
-        Create a cylinder, translate it, and use iterative closest point to
-        align mesh to its original position.
-
-        >>> import pyvista as pv
-        >>> import numpy as np
-        >>> source = pv.Cylinder(resolution=30).triangulate().subdivide(1)
-        >>> transformed = source.rotate_y(20).translate([-0.75, -0.5, 0.5])
-        >>> aligned = transformed.iterative_closest_point(source)
-        >>> _, closest_points = aligned.find_closest_cell(
-        ...     source.points, return_closest_point=True
-        ... )
-        >>> dist = np.linalg.norm(source.points - closest_points, axis=1)
-        >>> np.abs(dist).mean()  # doctest:+SKIP
-        9.997635192915073e-05
-
-        Visualize the source, transformed, and aligned meshes.
-
-        >>> pl = pv.Plotter(shape=(1, 2))
-        >>> _ = pl.add_text('Before Alignment')
-        >>> _ = pl.add_mesh(
-        ...     source, style='wireframe', opacity=0.5, line_width=2
-        ... )
-        >>> _ = pl.add_mesh(transformed)
-        >>> pl.subplot(0, 1)
-        >>> _ = pl.add_text('After Alignment')
-        >>> _ = pl.add_mesh(
-        ...     source, style='wireframe', opacity=0.5, line_width=2
-        ... )
-        >>> _ = pl.add_mesh(
-        ...     aligned,
-        ...     scalars=dist,
-        ...     scalar_bar_args={
-        ...         'title': 'Distance to Source',
-        ...         'fmt': '%.1E',
-        ...     },
-        ... )
-        >>> pl.show()
-
-        """
-        icp = _vtk.vtkIterativeClosestPointTransform()
-        icp.SetSource(self)
-        icp.SetTarget(target)
-        icp.GetLandmarkTransform().SetModeToRigidBody()
-        icp.SetMaximumNumberOfLandmarks(max_landmarks)
-        icp.SetMaximumMeanDistance(max_mean_distance)
-        icp.SetMaximumNumberOfIterations(max_iterations)
-        icp.SetCheckMeanDistance(check_mean_distance)
-        icp.SetStartByMatchingCentroids(start_by_matching_centroids)
-        icp.Update()
-        matrix = icp.GetMatrix()
-        if return_matrix:
-            return self.transform(matrix, inplace=False), matrix
-        return self.transform(matrix, inplace=False)
-
     def reconstruct_surface(self, nbr_sz=None, sample_spacing=None, progress_bar=False):
         """Reconstruct a surface from the points in this dataset.
