+++ tests/filters/test_poly_data.py
@@ -1,4 +1,3 @@
-import numpy as np
 import pytest
 
 import pyvista as pv
@@ -46,19 +45,3 @@ def test_boolean_intersect_edge_case():
 
     with pytest.warns(UserWarning, match='contained within another'):
         a.boolean_intersection(b)
-
-
-def test_iterative_closest_point():
-    # Create a simple mesh
-    source = pv.Cylinder(resolution=30).triangulate().subdivide(1)
-    transformed = source.rotate_y(20).rotate_z(25).translate([-0.75, -0.5, 0.5])
-
-    # Perform ICP registration
-    aligned = transformed.iterative_closest_point(source)
-
-    # Check if the number of points in the aligned mesh is the same as the original mesh
-    assert source.n_points == aligned.n_points
-
-    _, closest_points = aligned.find_closest_cell(source.points, return_closest_point=True)
-    dist = np.linalg.norm(source.points - closest_points, axis=1)
-    assert np.abs(dist).mean() < 1e-3
